Proyectos Treinta/bitrix_leads_fb/pyproject.toml
[tool.poetry]
name = "bitrix-leads-fb"
version = "0.1.0"
description = ""
authors = ["stivy2 <steven.porras@treinta.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fast-bitrix24 = "^1.5.16"
google-cloud-bigquery = "^3.11.4"
google-auth = "^2.22.0"
db-dtypes = "^1.1.1"
gspread = "^5.10.0"
oauth2client = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
